Bugfix 3.2.67

+++ tools/sendEmailAsync.sh
@@ -19,40 +19,46 @@ mkdir -p $mail_archiv_path
 chown gisadmin.www-data $mail_archiv_path
 chmod g+w $mail_archiv_path
 
-
-file=`find $mail_queue_path -name "email*" | head -n 1`
-
-while [ ! -z $file -a -e $file ]
-do
-  #echo "file: $file exists."
-  to_email=`cat $file | jq -r '.to_email'`
-  from_email=`cat $file | jq -r '.from_email'`
-  subject=`cat $file | jq -r '.subject'`
-  message=`cat $file | jq -r '.message'`
-  attachment=`cat $file | jq -r '.attachment'`
-
-  #tls=auto will only use tls if available
-  if [[ -z $attachment ]]; then
-    #echo Ohne Attachement sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=yes -u ${subject} -m ${message}
-    sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=auto -xu ${mail_smtp_user} -xp ${mail_smtp_password} -o message-charset=utf8 -u "${subject}" -m "${message}" > /dev/null 2>&1
-    #sendEmail -v -t 'peter.korduan@gdi-service.de' -f 'info@gdi-service.de' -s smtp.ionos.de:587 -o tls=yes -xu 'peter.korduan@gdi-backup.de' -xp '*****' -o message-charset=utf8 -u "Testkvwmap" -m "TestMessage"
-  else
-    #echo Mit attachement sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=yes -u ${subject} -m ${message} -a $attachment
-    #sendEmail -v -t $to_email -f $from_email -s ${smtp} -o tls=yes  -xu ${mail_smtp_user} -xp ${mail_smtp_password} -o message-charset=utf8 -u "TestPlandigital" -m "Testcontent" -a $attachment > /dev/null 2>&1
-    sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=auto -u "${subject}" -m "${message}" -xu ${mail_smtp_user} -xp ${mail_smtp_password} -o message-charset=utf8
-    if [[ -z $mail_copy_attachment ]]; then
-  
-      mv $attachment $mail_archiv_path
-    
-    else 
-      if [[ -z $attachment ]]; then
-        cp $attachment $mail_archiv_path
-      fi
-    
-    fi
-  fi
-  mv $file $mail_archiv_path
-  #echo "E-Mail $file gesendet."
+busyfile="${MAILQUEUEPATH}/busyfile"
+if [ -f "${busyfile}" ]; then
+  # Es gibt schon einen Prozess der die E-Mail-Dateien abarbeitet.
+  echo "${busyfile} exists."
+else
   file=`find $mail_queue_path -name "email*" | head -n 1`
-done
-#echo 'fertig'
+
+  while [ ! -z $file -a -e $file ]
+  do
+    #echo "file: $file exists."
+    to_email=`cat $file | jq -r '.to_email'`
+    from_email=`cat $file | jq -r '.from_email'`
+    subject=`cat $file | jq -r '.subject'`
+    message=`cat $file | jq -r '.message'`
+    attachment=`cat $file | jq -r '.attachment'`
+    attachment=${attachement/var/home/gisadmin}
+
+    #tls=auto will only use tls if available
+    if [[ -z $attachment ]]; then
+      #echo Ohne Attachement sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=yes -u ${subject} -m ${message}
+      sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=auto -xu ${mail_smtp_user} -xp ${mail_smtp_password} -o message-charset=utf8 -u "${subject}" -m "${message}" > /dev/null 2>&1
+      #sendEmail -v -t 'peter.korduan@gdi-service.de' -f 'info@gdi-service.de' -s smtp.ionos.de:587 -o tls=yes -xu 'peter.korduan@gdi-backup.de' -xp '*****' -o message-charset=utf8 -u "Testkvwmap" -m "TestMessage"
+    else
+      #echo Mit attachement sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=yes -u ${subject} -m ${message} -a $attachment
+      #sendEmail -v -t $to_email -f $from_email -s ${smtp} -o tls=yes  -xu ${mail_smtp_user} -xp ${mail_smtp_password} -o message-charset=utf8 -u "TestPlandigital" -m "Testcontent" -a $attachment > /dev/null 2>&1
+      sendEmail -v -t $to_email -f $from_email -s ${smtp_server}:${smtp_port} -o tls=auto -u "${subject}" -m "${message}" -xu ${mail_smtp_user} -xp ${mail_smtp_password} -o message-charset=utf8 -a $attachment > /dev/null 2>&1
+      if [[ -z $mail_copy_attachment ]]; then
+          mv $attachment $mail_archiv_path
+      else 
+        if [[ -z $attachment ]]; then
+          cp $attachment $mail_archiv_path
+        fi
+      fi
+    fi
+    mv $file $mail_archiv_path
+    #echo "E-Mail $file gesendet."
+    file=`find $mail_queue_path -name "email*" | head -n 1` # Frage die nächste Mail-Datei aus dem Stapel ab.
+  done
+  # Keine Mail-Dateien mehr im Stapel, lösche busyfile
+  if [ -f "${busyfile}" ]; then
+    rm $busyfile
+  fi
+fi
